requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pbr==3.1.1
python-decouple==3.1
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
